// src/views/mentor/entrevistas/entrevistasPendientes/Types.ts
export interface ListaEntrevistas {
    idEntrevistaEstudiante: number;
    nombreEstudiante: string;
    notaFinal: number;
}

export interface PaginatedResponse {
    content: ListaEntrevistas[];
    pageable: {
        pageNumber: number;
        pageSize: number;
        sort: {
            empty: boolean;
            sorted: boolean;
            unsorted: boolean;
        };
        offset: number;
        paged: boolean;
        unpaged: boolean;
    };
    last: boolean;
    totalElements: number;
    totalPages: number;
    size: number;
    number: number;
    sort: {
        empty: boolean;
        sorted: boolean;
        unsorted: boolean;
    };
    first: boolean;
    numberOfElements: number;
    empty: boolean;
}